requirements.txt
python-gflags==3.1.1
kubernetes==9.0.0
psycopg2-binary==2.9.8
Jinja2==3.1.3